requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
pandas
numpy
selenium
webdriver-manager
openpyxl
scikit-learn